# Avoid transcoding allocations when serializing doubles