--- a/libs/common/RespWriteUtils.cs
+++ b/libs/common/RespWriteUtils.cs
@@ -443,8 +443,6 @@ namespace Garnet.common
             if (double.IsInfinity(value))
             {
                 var infinityOutput = new Span<byte>(curr, (int)(end - curr));
-                curr += 10;
-
                 if (double.IsPositiveInfinity(value))
                 {
                     if (!"$4\r\n+inf\r\n"u8.TryCopyTo(infinityOutput))
@@ -456,6 +454,7 @@ namespace Garnet.common
                         return false;
                 }
 
+                curr += 10;
                 return true;
             }