--- a/libs/server/Objects/SortedSetGeo/SortedSetGeoObjectImpl.cs
+++ b/libs/server/Objects/SortedSetGeo/SortedSetGeoObjectImpl.cs
@@ -25,7 +25,7 @@ namespace Garnet.server
             public Byte[] Member;
             public double Distance;
             public string GeoHashCode;
-            public (double, double) Coords;
+            public (double Latitude, double Longitude) Coordinates;
         }
 
         /// <summary>
@@ -250,7 +250,7 @@ namespace Garnet.server
                                         distance :
                                         server.GeoHash.ConvertMetersToUnits(distance, units);
 
-                    while (!RespWriteUtils.WriteAsciiBulkString(distanceValue.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                    while (!RespWriteUtils.WriteDoubleBulkString(distanceValue, ref curr, end))
                         ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
                     countDone = count;
@@ -331,10 +331,10 @@ namespace Garnet.server
                         while (!RespWriteUtils.WriteArrayLength(2, ref curr, end))
                             ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
-                        while (!RespWriteUtils.WriteAsciiBulkString(lon.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                        while (!RespWriteUtils.WriteDoubleBulkString(lon, ref curr, end))
                             ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
-                        while (!RespWriteUtils.WriteAsciiBulkString(lat.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                        while (!RespWriteUtils.WriteDoubleBulkString(lat, ref curr, end))
                             ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
                     }
                     else
@@ -499,7 +499,7 @@ namespace Garnet.server
                                         Member = point.Item2,
                                         Distance = distance,
                                         GeoHashCode = server.GeoHash.GetGeoHashCode((long)point.Item1),
-                                        Coords = server.GeoHash.GetCoordinatesFromLong((long)point.Item1)
+                                        Coordinates = server.GeoHash.GetCoordinatesFromLong((long)point.Item1)
                                     });
 
                                     if (opts.WithCount && responseData.Count == countValue)
@@ -520,17 +520,17 @@ namespace Garnet.server
                             var distanceValue = (byBoxUnits.Length == 1 && (byBoxUnits[0] == (int)'M' || byBoxUnits[0] == (int)'m')) ? item.Distance
                                                 : server.GeoHash.ConvertMetersToUnits(item.Distance, byBoxUnits);
 
-                            while (!RespWriteUtils.WriteAsciiBulkString(distanceValue.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                            while (!RespWriteUtils.WriteDoubleBulkString(distanceValue, ref curr, end))
                                 ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
                             // Write array of 2 values
                             while (!RespWriteUtils.WriteArrayLength(2, ref curr, end))
                                 ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
-                            while (!RespWriteUtils.WriteAsciiBulkString(item.Coords.Item2.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                            while (!RespWriteUtils.WriteDoubleBulkString(item.Coordinates.Longitude, ref curr, end))
                                 ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
 
-                            while (!RespWriteUtils.WriteAsciiBulkString(item.Coords.Item1.ToString(CultureInfo.InvariantCulture), ref curr, end))
+                            while (!RespWriteUtils.WriteDoubleBulkString(item.Coordinates.Latitude, ref curr, end))
                                 ObjectUtils.ReallocateOutput(ref output, ref isMemory, ref ptr, ref ptrHandle, ref curr, ref end);
                         }