--- a/libs/server/Resp/Objects/SortedSetCommands.cs
+++ b/libs/server/Resp/Objects/SortedSetCommands.cs
@@ -1226,14 +1226,14 @@ namespace Garnet.server
 
                     if (result != null)
                     {
-                        foreach (var item in result)
+                        foreach (var (element, score) in result)
                         {
-                            while (!RespWriteUtils.WriteBulkString(item.Key, ref dcurr, dend))
+                            while (!RespWriteUtils.WriteBulkString(element, ref dcurr, dend))
                                 SendAndReset();
 
                             if (withscoresInclude)
                             {
-                                while (!RespWriteUtils.WriteAsciiBulkString(item.Value.ToString(CultureInfo.InvariantCulture), ref dcurr, dend))
+                                while (!RespWriteUtils.WriteDoubleBulkString(score, ref dcurr, dend))
                                     SendAndReset();
                             }
                         }