@badrishc 2024-05-30
Great to see the allocation avoidance, seems there might be several other places this can be used.